Master merge lynx 12.8.0

+++ alpha/lib/model/LiveEntryServerNode.php
@@ -155,7 +155,8 @@ class LiveEntryServerNode extends EntryServerNode
 		}
 		
 		/* @var $liveEntry LiveEntry */
-		if($this->getDc() === kDataCenterMgr::getCurrentDcId() && !$liveEntry->isCacheValid($this))
+		$timeFromLastUpdate = time() - $this->getUpdatedAt(null);
+		if($this->getDc() === kDataCenterMgr::getCurrentDcId() && !$liveEntry->isCacheValid($this) && $timeFromLastUpdate > LiveEntry::DEFAULT_CACHE_EXPIRY)
 		{
 			KalturaLog::info("Removing media server id [" . $this->getServerNodeId() . "] from liveEntry [" . $this->getEntryId() . "]");
 			$this->deleteOrMarkForDeletion($liveEntry);
